move svg parser to a more reasonable location there is still some work to be done here
